Move CommentSkipIterRev to byte_iter
